requirements.txt
art==6.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.6
loguru==0.7.2
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
termcolor==2.4.0
urllib3==2.2.0
win32-setctime==1.1.0
wonderwords==2.2.0
